New dataloader & fix circulars
